fix homeserver and start xbps

+++ justfile
@@ -78,5 +78,5 @@ NIXNAME := "vm-aarch64"
 
   ! [ -d ~/.config/nushell ] && ln -sf /nix-config/mut/nushell ~/.config/nushell
   ! [ -d ~/.config/vis ] && ln -sf /nix-config/mut/vis ~/.config/vis
-  rm -rf "$HOME/Library/Application Support/nushell";  ln -sf /nix-config/mut/nushell  "$HOME/Library/Application Support/nushell"
+  [ -d "$HOME/Library/Application Support" ] && { rm -rf "$HOME/Library/Application Support/nushell";  ln -sf /nix-config/mut/nushell  "$HOME/Library/Application Support/nushell"; }
   true
